ImageSteganographySystem-1/requirements.txt
flask==2.0.1
pillow==9.0.0
werkzeug==2.0.1
numpy==1.21.0
scikit-image==0.18.3
opencv-python==4.5.3.56
pefile==2021.9.3
yara-python==4.1.0 